requirements.txt
google-cloud-storage
s3-client
pytest
coverage
